Define policies to control how pool types and names are defined
