+++ immer/extra/persist/json/persistable.hpp
@@ -2,7 +2,7 @@
 
 #include <immer/extra/persist/errors.hpp>
 #include <immer/extra/persist/json/json_immer.hpp>
-#include <immer/extra/persist/json/json_with_pool.hpp>
+#include <immer/extra/persist/json/pools.hpp>
 #include <immer/extra/persist/traits.hpp>
 
 #include <fmt/format.h>
@@ -69,20 +69,20 @@ struct persistable
 
 template <class Previous,
           class Storage,
-          class Names,
-          class WrapF,
+          class WrapFn,
+          class PoolNameFn,
           class Container>
-auto save_minimal(
-    const json_immer_output_archive<Previous,
-                                    detail::output_pools<Storage, Names>,
-                                    WrapF>& ar,
-    const persistable<Container>& value)
+auto save_minimal(const json_immer_output_archive<Previous,
+                                                  detail::output_pools<Storage>,
+                                                  WrapFn,
+                                                  PoolNameFn>& ar,
+                  const persistable<Container>& value)
 {
     auto& pool =
-        const_cast<
-            json_immer_output_archive<Previous,
-                                      detail::output_pools<Storage, Names>,
-                                      WrapF>&>(ar)
+        const_cast<json_immer_output_archive<Previous,
+                                             detail::output_pools<Storage>,
+                                             WrapFn,
+                                             PoolNameFn>&>(ar)
             .get_output_pools()
             .template get_output_pool<Container>();
     auto [pool2, id] = add_to_pool(value.container, std::move(pool));
@@ -94,27 +94,32 @@ auto save_minimal(
 // possible to have only load_minimal for a type without having save_minimal.
 template <class Previous,
           class Storage,
-          class Names,
-          class WrapF,
+          class WrapFn,
+          class PoolNameFn,
           class Container>
-auto save_minimal(
-    const json_immer_output_archive<Previous,
-                                    detail::input_pools<Storage, Names>,
-                                    WrapF>& ar,
-    const persistable<Container>& value) ->
+auto save_minimal(const json_immer_output_archive<Previous,
+                                                  detail::input_pools<Storage>,
+                                                  WrapFn,
+                                                  PoolNameFn>& ar,
+                  const persistable<Container>& value) ->
     typename container_traits<Container>::container_id::rep_t
 {
     throw std::logic_error{"Should never be called"};
 }
 
-template <class Previous, class Pools, class WrapF, class Container>
+template <class Previous,
+          class Pools,
+          class WrapFn,
+          class PoolNameFn,
+          class Container>
 void load_minimal(
-    const json_immer_input_archive<Previous, Pools, WrapF>& ar,
+    const json_immer_input_archive<Previous, Pools, WrapFn, PoolNameFn>& ar,
     persistable<Container>& value,
     const typename container_traits<Container>::container_id::rep_t& id)
 {
     auto& loader =
-        const_cast<json_immer_input_archive<Previous, Pools, WrapF>&>(ar)
+        const_cast<
+            json_immer_input_archive<Previous, Pools, WrapFn, PoolNameFn>&>(ar)
             .template get_loader<Container>();
 
     // Have to be specific because for vectors container_id is different from
